ODA-374 Looking to the results of the reaction_advection model we noticed a strange periodical drop in standard deviation of the ensemble.
The problem is caused by a rounding issue in 
private void propagateNoiseModelsAndAddNoiseToExchangeItems(ITime currentTime, ITime targetTime) 

In matching the model times to noise model times we have a test

if(time>this.lastNoiseTimes.get(modelExchangeItemId)){

When the noise model and model make the same time steps, this test is very likely to give an unwanted result due to timy rounding errors. Is should be changed into 

if(time>this.lastNoiseTimes.get(modelExchangeItemId)+0.5*timePrecision){

Taking roudning errors into account. The factor 0.5 is used as an additional safeguard since times are compared (using timePrecision) within the if block.