ODA-419
Extended interface for (temporarely) handling localization (Cohn) combined with assimilating the observations one at a time. 
NOTE: the weights in the observer SHOULD corresponds to those in the model.
This is directly why this setup is far from ideal, will probably be removed in the future when we reimplement localization in OpenDA!

Former-commit-id: e8924ca583739b98a227c5d323bf6c8996f1433f