#57 Fixed the tests
D3dWindMask files failed due to an incomplete introduction of a new 2d-grid-field class. Has for now be refactored as extension f the existing 2d-field. Should in the end completely replace that class.
The NetcdfD3d-his file failed because the test was not updated according to an extension in the implementation (new exchange items) and to an new test file (17 instead of 13 time steps).