Add MPI init, finialise, and subroutine calls
This propagates the MPI patches into the OpenDA wrapper. The MPI
environment is initialised when the wrapper is set up. The domain
decomposition seems required before the second model init, as is done in
the EFDC.for main file. When the wrapper is destroyed, the MPI
environment is destroyed as well.

NOTE: This *ONLY* considers running with at most 1 rank and no care is
taken to restrict certain operations to only the first rank, as would be
needed for file IOa and/or log printing.