Use Unicode paths for loading DLLs on Windows
Without the UNICODE define, the Windows API will use ANSI strings
instead of Unicode strings. The implementation in Costa assumes that
Unicode strings are used, and defines the library name as a 16-bit
WCHAR. Copying over a 8-bit char string to a 16-bit WCHAR string
effectively result in a \0 after that first character, hence e.g.
LoadLibrary will attempt to load a DLL with a name consisting of only
the first character of the string...